Split UpdateMonitor from status
